update 27 October 11:22

+++ routes/web.php
@@ -25,10 +25,7 @@ use App\Http\Controllers\HabilitationAdminController;
 |
 */
 Auth::routes();
-Route::get('/', function () {
-    return redirect()->route('login');
-});
-
+Route::view('', 'auth.login')->name('/');
 Route::get('/home', [HomeController::class, 'index'])->name('home');
 
 /* USUARIO */
@@ -48,75 +45,63 @@ Route::post('updatepass', [UserController::class, 'updatepass'])->name('user.upd
 Route::get('profile',[HomeController::class, 'profile'])->name('profile');
 
 /* MODULOS */
-Route::get('/modulo1/lista', [ConformidadObraController::class, 'index'])->name('conformidad.index');
+/* Route::get('/modulo1/lista', [ConformidadObraController::class, 'index'])->name('conformidad.index');
 Route::get('/modulo1/form', [ConformidadObraController::class, 'create'])->name('conformidad.create');
 Route::post('/conformidad/obra', [ConformidadObraController::class, 'store'])->name('conformidad.store');
 Route::get('/modulo1/edit/{id}', [ConformidadObraController::class, 'edit'])->name('conformidad.edit');
 Route::post('/modulo1/update/{id}', [ConformidadObraController::class, 'update'])->name('conformidad.update');
-Route::delete('/modulo1/delete/{id}', [ConformidadObraController::class, 'destroy'])->name('conformidad.destroy');
+Route::delete('/modulo1/delete/{id}', [ConformidadObraController::class, 'destroy'])->name('conformidad.destroy'); */
 
-/* Route::get('/pisos/lista/', [PisoController::class, 'index'])->name('piso.index');
-Route::get('/pisos/create/', [PisoController::class, 'create'])->name('piso.create');
-Route::post('/pisos/store/', [PisoController::class, 'store'])->name('piso.store');
-Route::get('/pisos/edit/{id}', [ConformidadObraController::class, 'edit'])->name('piso.edit');
-Route::post('/pisos/update/{id}', [ConformidadObraController::class, 'update'])->name('piso.update');
-Route::delete('/pisos/delete/{id}', [ConformidadObraController::class, 'destroy'])->name('piso.destroy');
- */
-
-Route::get('/modulo2/lista', [HabilitacionUrbController::class, 'index'])->name('habilitacion.index');
+/* Route::get('/modulo2/lista', [HabilitacionUrbController::class, 'index'])->name('habilitacion.index');
 Route::get('/modulo2/form', [HabilitacionUrbController::class, 'create'])->name('habilitacion.create');
 Route::post('/habilitacion/urb', [HabilitacionUrbController::class, 'store'])->name('habilitacion.store');
 Route::get('/modulo2/edit/{id}', [HabilitacionUrbController::class, 'edit'])->name('habilitacion.edit');
 Route::post('/modulo2/update/{id}', [HabilitacionUrbController::class, 'update'])->name('habilitacion.update');
-Route::delete('/modulo2/delete/{id}', [HabilitacionUrbController::class, 'destroy'])->name('habilitacion.destroy');
+Route::delete('/modulo2/delete/{id}', [HabilitacionUrbController::class, 'destroy'])->name('habilitacion.destroy'); */
 
-Route::get('/modulo3/lista', [ConstanciaPosesionController::class, 'index'])->name('constancia.index');
-Route::get('/modulo3/form', [ConstanciaPosesionController::class, 'create'])->name('constancia.create');
-Route::post('/constancia/posesion', [ConstanciaPosesionController::class, 'store'])->name('constancia.store');
-Route::get('/modulo3/edit/{id}', [ConstanciaPosesionController::class, 'edit'])->name('constancia.edit');
-Route::post('/modulo3/update/{id}', [ConstanciaPosesionController::class, 'update'])->name('constancia.update');
-Route::delete('/modulo3/delete/{id}', [ConstanciaPosesionController::class, 'destroy'])->name('constancia.destroy');
+Route::get('/modulo3/lista', [ConstanciaPosesionController::class, 'index'])->name('constancia.index')->middleware('role:user,admin,root');
+Route::get('/modulo3/form', [ConstanciaPosesionController::class, 'create'])->name('constancia.create')->middleware('role:user,admin,root');
+Route::post('/constancia/posesion', [ConstanciaPosesionController::class, 'store'])->name('constancia.store')->middleware('role:user,admin,root');
+Route::get('/modulo3/edit/{id}', [ConstanciaPosesionController::class, 'edit'])->name('constancia.edit')->middleware('role:user,admin,root');
+Route::post('/modulo3/update/{id}', [ConstanciaPosesionController::class, 'update'])->name('constancia.update')->middleware('role:user,admin,root');
+Route::delete('/modulo3/delete/{id}', [ConstanciaPosesionController::class, 'destroy'])->name('constancia.destroy')->middleware('role:user,admin,root');
 
-Route::get('/modulo4/lista', [ParametrosUrbController::class, 'index'])->name('parametro.index');
+/* Route::get('/modulo4/lista', [ParametrosUrbController::class, 'index'])->name('parametro.index');
 Route::get('/modulo4/form', [ParametrosUrbController::class, 'create'])->name('parametro.create');
 Route::post('/parametro/urbano', [ParametrosUrbController::class, 'store'])->name('parametro.store');
 Route::get('/modulo4/edit/{id}', [ParametrosUrbController::class, 'edit'])->name('parametro.edit');
 Route::post('/modulo4/update/{id}', [ParametrosUrbController::class, 'update'])->name('parametro.update');
-Route::delete('/modulo4/delete/{id}', [ParametrosUrbController::class, 'destroy'])->name('parametro.destroy');
+Route::delete('/modulo4/delete/{id}', [ParametrosUrbController::class, 'destroy'])->name('parametro.destroy'); */
 
-Route::get('/modulo5/lista', [TrabViaPublicaController::class, 'index'])->name('via.pub.index');
+/* Route::get('/modulo5/lista', [TrabViaPublicaController::class, 'index'])->name('via.pub.index');
 Route::get('/modulo5/form', [TrabViaPublicaController::class, 'create'])->name('via.pub.create');
 Route::post('/trab/viapublica', [TrabViaPublicaController::class, 'store'])->name('via.pub.store');
 Route::get('/modulo5/edit/{id}', [TrabViaPublicaController::class, 'edit'])->name('via.pub.edit');
 Route::post('/modulo5/update/{id}', [TrabViaPublicaController::class, 'update'])->name('via.pub.update');
-Route::delete('/modulo5/delete/{id}', [TrabViaPublicaController::class, 'destroy'])->name('via.pub.destroy');
+Route::delete('/modulo5/delete/{id}', [TrabViaPublicaController::class, 'destroy'])->name('via.pub.destroy'); */
 
 /* PDF'S */
 Route::get('co/fpdf/{token}', [ConformidadObraController::class, 'pdf'])->name('pdf.co');
 Route::get('hu/fpdf/{token}', [HabilitacionUrbController::class, 'pdf'])->name('pdf.hu');
-
 Route::get('ConstanciaPosesion/{id}', [ConstanciaPosesionController::class, 'pdf'])->name('pdf.cp');
 
-
-Route::get('controlEstImp', [HabilitationAdminController::class, 'index'])->name('administrador.index');
-Route::post('unlockPrint/{id}', [HabilitationAdminController::class, 'unlockPrint'])->name('administrador.unlockprint');
-Route::get('lockPrint/{id}', [HabilitationAdminController::class, 'lokckPrint'])->name('administrador.lockprint');
-Route::post('disable/{id}', [HabilitationAdminController::class, 'disable'])->name('administrador.disable');
-Route::post('enable/{id}', [HabilitationAdminController::class, 'enable'])->name('administrador.enable');
+Route::get('controlEstImp', [HabilitationAdminController::class, 'index'])->name('administrador.index')->middleware('role:admin,root');
+Route::post('unlockPrint/{id}', [HabilitationAdminController::class, 'unlockPrint'])->name('administrador.unlockprint')->middleware('role:admin,root');
+Route::get('lockPrint/{id}', [HabilitationAdminController::class, 'lokckPrint'])->name('administrador.lockprint')->middleware('role:admin,root');
+Route::post('disable/{id}', [HabilitationAdminController::class, 'disable'])->name('administrador.disable')->middleware('role:admin,root');
+Route::post('enable/{id}', [HabilitationAdminController::class, 'enable'])->name('administrador.enable')->middleware('role:admin,root');
 
 Route::get('pu/fpdf/{token}', [ParametrosUrbController::class, 'pdf'])->name('pdf.pu');
 Route::get('tvp/fpdf/{token}', [TrabViaPublicaController::class, 'pdf'])->name('pdf.tvp');
 
 Route::get('actualizar/pisos/{id}/{pisos}', [ConformidadObraController::class, 'actualizarPisos'])->name('actualizar.pisos');
-/* 
-Route::post('/actualizar/pisos/{id}/{pisos}', [ConformidadObraController::class, 'actualizarCantidadPisos'])->name('actualizar.pisos'); */
-
+/* Route::post('/actualizar/pisos/{id}/{pisos}', [ConformidadObraController::class, 'actualizarCantidadPisos'])->name('actualizar.pisos'); */
 Route::get('licencias/getSunatDatos/{ruc}', [ParametrosUrbController::class, 'getSunatDatos'])->name('getSunatDatos');
 
 /* UBICACION */
-Route::get('ubicacion', [UbicacionController::class, 'index'])->name('ubicacion.index');
-Route::post('ubicacion/create', [UbicacionController::class, 'store'])->name('ubicacion.store');
-Route::post('ubicacion/update/{id}', [UbicacionController::class, 'update'])->name('ubicacion.update');
-Route::delete('ubicacion/destroy/{id}', [UbicacionController::class, 'destroy'])->name('ubicacion.destroy');
+Route::get('ubicacion', [UbicacionController::class, 'index'])->name('ubicacion.index')->middleware('role:admin,root');
+Route::post('ubicacion/create', [UbicacionController::class, 'store'])->name('ubicacion.store')->middleware('role:admin,root');
+Route::post('ubicacion/update/{id}', [UbicacionController::class, 'update'])->name('ubicacion.update')->middleware('role:admin,root');
+Route::delete('ubicacion/destroy/{id}', [UbicacionController::class, 'destroy'])->name('ubicacion.destroy')->middleware('role:admin,root');
 
 Route::get('ubicacion/getSector/{zona}', [UbicacionController::class, 'getSector'])->name('getSector');
